Set recursion limit to 8 and create integration tests directory for xcm-executor
