+++ xcm/xcm-executor/src/lib.rs
@@ -42,7 +42,7 @@ pub use config::Config;
 pub struct XcmExecutor<Config>(PhantomData<Config>);
 
 /// The maximum recursion limit for `execute_xcm` and `execute_effects`.
-pub const MAX_RECURSION_LIMIT: u32 = 80;
+pub const MAX_RECURSION_LIMIT: u32 = 8;
 
 impl<Config: config::Config> ExecuteXcm<Config::Call> for XcmExecutor<Config> {
 	fn execute_xcm_in_credit(
@@ -354,132 +354,5 @@ impl<Config: config::Config> XcmExecutor<Config> {
 			_ => return Err(XcmError::UnhandledEffect)?,
 		}
 		Ok(total_surplus)
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	use polkadot_test_client::{
-		BlockBuilderExt, ClientBlockImportExt, DefaultTestClientBuilderExt, ExecutionStrategy,
-		InitPolkadotBlockBuilder, TestClientBuilder, TestClientBuilderExt,
-	};
-	use polkadot_test_runtime::new_test_ext;
-	use polkadot_test_service::construct_extrinsic;
-	use sp_runtime::{generic::BlockId, traits::Block};
-	use sp_state_machine::InspectState;
-	use xcm::v0::prelude::*;
-
-	#[test]
-	fn execute_within_recursion_limit() {
-		new_test_ext().execute_with(|| {
-			let mut client = TestClientBuilder::new()
-				.set_execution_strategy(ExecutionStrategy::AlwaysWasm)
-				.build();
-
-			let mut msg = WithdrawAsset {
-				assets: vec![ConcreteFungible { id: MultiLocation::Null, amount: 0 }],
-				effects: vec![],
-			};
-			for _ in 0..MAX_RECURSION_LIMIT {
-				msg = WithdrawAsset {
-					assets: vec![ConcreteFungible { id: MultiLocation::Null, amount: 0 }],
-					effects: vec![BuyExecution {
-						fees: All,
-						weight: 0,
-						debt: 0,
-						halt_on_error: true,
-						// nest `msg` into itself on each iteration.
-						xcm: vec![msg],
-					}],
-				};
-			}
-
-			let execute = construct_extrinsic(
-				&client,
-				polkadot_test_runtime::Call::Xcm(pallet_xcm::Call::execute(
-					Box::new(msg.clone()),
-					1_000_000_000,
-				)),
-				sp_keyring::Sr25519Keyring::Alice,
-			);
-
-			let mut block_builder = client.init_polkadot_block_builder();
-			block_builder.push_polkadot_extrinsic(execute).expect("pushes extrinsic");
-
-			let block = block_builder.build().expect("Finalizes the block").block;
-			let block_hash = block.hash();
-
-			futures::executor::block_on(client.import(sp_consensus::BlockOrigin::Own, block))
-				.expect("imports the block");
-
-			client
-				.state_at(&BlockId::Hash(block_hash))
-				.expect("state should exist")
-				.inspect_state(|| {
-					assert!(polkadot_test_runtime::System::events().iter().any(|r| matches!(
-						r.event,
-						polkadot_test_runtime::Event::Xcm(pallet_xcm::Event::Attempted(
-							Outcome::Complete(_)
-						),),
-					)));
-				});
-		});
-	}
-
-	#[test]
-	fn exceed_recursion_limit() {
-		let mut client = TestClientBuilder::new()
-			.set_execution_strategy(ExecutionStrategy::AlwaysWasm)
-			.build();
-
-		let mut msg = WithdrawAsset {
-			assets: vec![ConcreteFungible { id: MultiLocation::Null, amount: 0 }],
-			effects: vec![],
-		};
-		for _ in 0..(MAX_RECURSION_LIMIT + 1) {
-			msg = WithdrawAsset {
-				assets: vec![ConcreteFungible { id: MultiLocation::Null, amount: 0 }],
-				effects: vec![BuyExecution {
-					fees: All,
-					weight: 0,
-					debt: 0,
-					halt_on_error: true,
-					// nest `msg` into itself on each iteration.
-					xcm: vec![msg],
-				}],
-			};
-		}
-
-		let execute = construct_extrinsic(
-			&client,
-			polkadot_test_runtime::Call::Xcm(pallet_xcm::Call::execute(
-				Box::new(msg.clone()),
-				1_000_000_000,
-			)),
-			sp_keyring::Sr25519Keyring::Alice,
-		);
-
-		let mut block_builder = client.init_polkadot_block_builder();
-		block_builder.push_polkadot_extrinsic(execute).expect("pushes extrinsic");
-
-		let block = block_builder.build().expect("Finalizes the block").block;
-		let block_hash = block.hash();
-
-		futures::executor::block_on(client.import(sp_consensus::BlockOrigin::Own, block))
-			.expect("imports the block");
-
-		client
-			.state_at(&BlockId::Hash(block_hash))
-			.expect("state should exist")
-			.inspect_state(|| {
-				assert!(polkadot_test_runtime::System::events().iter().any(|r| matches!(
-					r.event,
-					polkadot_test_runtime::Event::Xcm(pallet_xcm::Event::Attempted(
-						Outcome::Incomplete(_, XcmError::RecursionLimitReached),
-					)),
-				)));
-			});
 	}
 }
